Bug 1661: Improve build speed with better caching strategy
Changed the way that the cache is located, so the same code can run on any build machine without any configuration.

Made work around for Bug 419 the only option, as we're not using the alternative anyway, and I don't want to commit completely untested code!

Changed caching algorithm, to improve build times.